feat(new-issuer): connect container to presenter
