--- packages/new-polymath-issuer/src/pages/DividendsDetails/index.tsx
@@ -1,3 +1,0 @@
-import { Container } from './Container';
-
-export const DividendsDetails = Container;
