+++ packages/new-polymath-issuer/src/pages/index.tsx
@@ -19,9 +19,8 @@ export { SecurityTokensIndexPage } from './SecurityTokensIndex';
 export { SecurityTokensReservePage } from './SecurityTokensReserve';
 export { SecurityTokensDividendsPage } from './SecurityTokensDividends';
 export { DividendsWizardPage } from './DividendsWizard';
-export { DividendDetailsPage } from './DividendDetails';
 export { NotFoundPage } from './NotFound';
-export { DividendsDetails } from './DividendsDetails';
+export { DividendDetailsPage } from './DividendDetails';
 
 interface StateProps {
   router: RootState['router'];
